feat(v1.4): WebUI enhancement — component split, auto-refresh, dark mode, mobile responsive
Phase 15: Component Split
- Split App.tsx from 1,555 lines to 42 lines (-97%)
- Extract 7 page components to pages/ directory
- Extract 5 shared components to components/ directory
- Extract auth context to context/auth.tsx
- Extract format utilities to utils/format.ts

Phase 16: Real-time Data
- Create useAutoRefresh hook with configurable intervals
- OverviewPage: 30s auto-refresh with loading indicator
- FeedbackPage: 15s auto-refresh for GFI monitoring
- GateMonitorPage: 30s auto-refresh for gate stats

Phase 17: Dark Mode
- Create ThemeProvider with system preference detection
- Manual theme toggle button in sidebar
- Theme persistence via localStorage
- Add [data-theme=dark] CSS variables

Phase 18: Mobile Responsive
- Add responsive styles for theme toggle button
- Existing media queries handle 1200px/1024px/768px/640px breakpoints

Phase 19: UX Polish
- Add refresh indicator with pulse animation
- Add theme toggle with Sun/Moon icons
