+++ packages/openclaw-plugin/ui/src/components/ErrorState.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export function ErrorState({ error }: { error: string }) {
+  return <div className="panel error">{error}</div>;
+}
+
