+++ packages/openclaw-plugin/ui/src/components/index.ts
@@ -1,0 +1,5 @@
+export { Loading } from './Loading';
+export { ErrorState } from './ErrorState';
+export { ProtectedRoute } from './ProtectedRoute';
+export { Shell } from './Shell';
+export { WorkspaceConfig } from './WorkspaceConfig';
