+++ packages/openclaw-plugin/ui/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { App } from './App';
+import App from './App';
 import './styles.css';
 
 const container = document.getElementById('root');
