+++ packages/openclaw-plugin/ui/src/utils/format.ts
@@ -1,0 +1,15 @@
+export function formatPercent(value: number): string {
+  return `${(value * 100).toFixed(1)}%`;
+}
+
+export function formatDate(value: string | null): string {
+  if (!value) return 'N/A';
+  return new Date(value).toLocaleString();
+}
+
+export function formatDuration(ms: number | null): string {
+  if (ms === null) return 'N/A';
+  if (ms < 1000) return `${ms}ms`;
+  if (ms < 60000) return `${(ms / 1000).toFixed(1)}s`;
+  return `${(ms / 60000).toFixed(1)}min`;
+}
